throughputmod: add reference time option
Via --reference-time (-r) it is now possible to GNUplot all connections
from one pcap file with the same time-offset. Until now (without -r) the
start-time for GNUplot graphs is the particular connection (time 0.0).
With -r, the reference time is now the first packet in the capture file.

This option makes it possible to draw several flows starting to
different times without align then on the time-axis.

Signed-off-by: Hagen Paul Pfeifer <hagen@jauu.net>