core: add new decoder functionality
Due to python-pcap removal the packet decoder was enforced to interpret
every packet as Ethernet - which will not be true in every case.

This patch check for some common types (Ethernet, Loopback, Linux) and
if not correct we enforce a Ethernet encoding - which will _probably_
not be correct, but it is better then nothing. A big warning message is
printed.

Signed-off-by: Hagen Paul Pfeifer <hagen@jauu.net>